Avoid loading non-media files to RS5K
When using this script with ReaSamplomatic5000, the script would choose any file from the folder where already loaded sample is. But, it would also load non-media files from that folder and that would silence ReaSamplomatic5000 and user would need to run the script again and again just to get to a valid media file. This is now fixed.